+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.5.0] - 2022-04-23
+### Added
+- `AddRange` extension for `HashSet<T>`
+- `RemoveRange` extension for `HashSet<T>` and `IList<T>`
+
 ## [2.4.0] - 2022-04-23
 ### Added
 - `Database` now requires a generic `IEquatable<FNVHash>` key type for identifying items stored